+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changes for developers
 * `getInvocationInfo` is now also available separately as PHP command.
+* Constants and properties will now also return their default values, if present.
 * `getInvocationInfoAt` will now return an `offset` rather than a `bufferPosition` to be consistent with the other commands.
 * `getResultingTypesAt` is now simply a convenience call to `deduceTypes` as its underlying code has been completely moved to PHP.
 * `deduceTypes` gained a new parameter `ignoreLastElement`, which does the same as the identically-named parameter for `getResultingTypesAt` does.